+++ api/src/main/java/com/github/skjolber/packing/api/ep/ExtremePoints.java
@@ -1,5 +1,6 @@
 package com.github.skjolber.packing.api.ep;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.github.skjolber.packing.api.Placement2D;
@@ -14,7 +15,7 @@ public interface ExtremePoints<P extends Placement2D, Point extends Point2D<P>> 
 
 	int getValueCount();
 
-	void reset(int dx, int dy, int dz);
+	void reset(BigDecimal dx, BigDecimal dy, BigDecimal dz);
 
 	void redo();
 }
